Never panic in cases where we have a ref not a ptr.
In this (temporary?) fork of bindgen, types are passed around with
some annotations to indicate whether they're a reference or a pointer.
In some cases we output Rust attributes to state which they are
(which is then used downstream by autocxx).

In other cases, we either:
* discarded those annotations where we felt they weren't useful; or
* panicked if we thought they were useful but didn't know how to
  handle them.

This latter strategy turns out to have been a mistake, because
there are arbitrary suchs situations in a C++ codebase, and it doesn't
match the strategy of autocxx where it converts as much of the
bindings as possible and just warns about problems with the rest.
All these call sites have now moved to the former strategy.